Update interamap configuration and refactor imports for consistency
