+++ sinks/interamap/src/url_to_qrcode.py
@@ -6,7 +6,7 @@ import os
 from PIL.ImageQt import QPixmap
 from PySide6.QtCore import Qt
 from PySide6.QtWidgets import QMainWindow, QLabel, QApplication, QVBoxLayout, QWidget
-from src.config import TERMINAL_QR_CODE
+from config import TERMINAL_QR_CODE
 
 
 def generate_qr_code(qr_code_url):
